feat(header): hide calendar button when server doesn't support MSC3030
RoomHeader gains show_calendar_btn_ (default false). The button is only
drawn, hit-tested, and included in layout space reservation when enabled.
Each shell's on_server_info_ready_ui_() / _onServerInfoReady calls
header->set_jump_to_date_enabled(server_info_.supports_msc3030) and
triggers a relayout. On servers that don't advertise MSC3030 the button
never appears; on those that do it appears once server info is fetched.

Also fix two broken Qt6 includes that caused build failures:
- canvas_qpainter.cpp: add QTextBlock (QTextDocument doesn't pull it in)
- host_qt.cpp: add QClipboard (QGuiApplication no longer does so transitively)

+++ ui/shared/tk/canvas_qpainter.cpp
@@ -11,6 +11,7 @@
 #include <QtGui/QPainterPath>
 #include <QtGui/QPen>
 #include <QtGui/QTextOption>
+#include <QtGui/QTextBlock>
 #include <QtGui/QTextDocument>
 #include <QtGui/QAbstractTextDocumentLayout>
 #include <QtGui/QPalette>
@@ -529,7 +530,7 @@ private:
         QString plain = doc_->toPlainText();
         if (qt_pos <= 0)
             return 0;
-        int clamped = std::min(qt_pos, plain.size());
+        int clamped = std::min(qt_pos, static_cast<int>(plain.size()));
         return plain.left(clamped).toUtf8().size();
     }
